Move delete button and profile stuff to profile activity
