+++ app/src/main/AndroidManifest.xml
@@ -12,6 +12,9 @@
         android:theme="@style/Theme.MusicMap"
         tools:targetApi="31">
         <activity
+            android:name=".ProfileActivity"
+            android:exported="false" />
+        <activity
             android:name=".screens.main.HomeActivity"
             android:exported="false" />
         <activity
