+++ app/src/main/java/com/example/musicmap/screens/ProfileActivity.java
@@ -1,0 +1,82 @@
+package com.example.musicmap.screens;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.widget.Button;
+import android.widget.TextView;
+
+import com.example.musicmap.R;
+import com.example.musicmap.screens.auth.AuthActivity;
+import com.example.musicmap.util.firebase.AuthSystem;
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+
+public class ProfileActivity extends AppCompatActivity implements FirebaseAuth.AuthStateListener {
+
+    private FirebaseAuth auth;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_profile);
+
+        FirebaseUser firebaseUser = auth.getCurrentUser();
+
+        TextView emailVerified = findViewById(R.id.emailVerified_textView);
+        TextView uuidText = findViewById(R.id.uuid_textView);
+        TextView emailText = findViewById(R.id.email_textView);
+        TextView usernameText = findViewById(R.id.username_textView);
+
+        if (firebaseUser != null) {
+            firebaseUser.reload().continueWithTask(task -> {
+                if (firebaseUser.isEmailVerified()) {
+                    emailVerified.setText(getString(R.string.email_verified));
+                } else {
+                    emailVerified.setText(R.string.email_not_verified);
+                }
+                uuidText.setText(firebaseUser.getUid());
+                emailText.setText(firebaseUser.getEmail());
+                usernameText.setText(firebaseUser.getDisplayName());
+                return null;
+            });
+        }
+
+        Button logoutButton = findViewById(R.id.logout_button);
+        logoutButton.setOnClickListener(view -> logout());
+
+        Button deleteAccountButton = findViewById(R.id.deleteAccount_button);
+        deleteAccountButton.setOnClickListener(view -> deleteAccount());
+    }
+
+    @Override
+    public void onAuthStateChanged(@NonNull FirebaseAuth firebaseAuth) {
+        auth = firebaseAuth;
+        FirebaseUser firebaseUser = auth.getCurrentUser();
+
+        if (firebaseUser == null) {
+            Intent authIntent = new Intent(this, AuthActivity.class);
+            authIntent.addFlags(Intent.FLAG_ACTIVITY_NO_HISTORY);
+            startActivity(authIntent);
+            finish();
+        }
+    }
+
+    public void logout() {
+        auth.signOut();
+    }
+
+    // TODO refactor, should be elsewhere
+    public void deleteAccount() {
+        AuthSystem.deleteUser();
+    }
+
+    @Override
+    public void onStop() {
+        super.onStop();
+        auth.removeAuthStateListener(this);
+    }
+}
