--- app/src/main/res/drawable/bottom_nav.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="@color/primary" />
-    <corners android:radius="20dp" />
-</shape>
