+++ app/src/main/res/drawable/primary_rounded.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape android:shape="rectangle"
+    xmlns:android="http://schemas.android.com/apk/res/android" >
+    <solid android:color="@color/primary" />
+    <corners android:radius="16dp" />
+</shape>
